Add files via upload
Comment updates in the HTML and JS code. This version requires the video files to be in sub-directory 'videos'. Clicking the 'choose file' button will display a browser.  If a file is selected other than in the videos folder, the user will need to type the full path in the prompt dialog.